requirements.txt
pandas
numpy
matplotlib
scikit-learn
seaborn
flask
flask_sqlalchemy
werkzeug
python-dotenv